traslado de card a main.js y agrega template a la tarjeta
