+++ src/index.html
@@ -5,6 +5,7 @@
   <meta charset="utf-8">
   <title>Pokémon Data</title>
   <link rel="stylesheet" href="style.css" />
+  <link rel="shortcut icon" href="img/favicon.png" type="image/x-icon">
   <link
     href="https://fonts.googleapis.com/css2?family=Iceland&family=Roboto:ital,wght@1,700&family=Space+Grotesk:wght@500&display=swap"
     rel="stylesheet">
@@ -61,7 +62,9 @@
         </p>
       </div>
     </section>
- 
+  <section class="counterContainer">
+      <p class="Pokemones"> "Pokemones (" <spam id="quantity"> 251 </spam> ")" </p>
+  </section>
     
 
   </main>
